// src/utils/enums/list.ts
/* eslint-disable no-unused-vars */
/**
 * Enum to define the direction of list
 * @enum {String}
 * @readonly
 */
export enum ListDirection {
  vertical = 'vertical',
  horizontal = 'horizontal'
}

export enum ListItemType {
  string = 'string',
  status = 'status',
  string_to_copy = 'string_to_copy',
  date = 'date',
  amount = 'amount'
}